Allow embeds to be instantiated in a sandboxed host which may be hosted in a looker instance.
The iframe source is now prefixed with the looker instance URL if the auth URL is not provided.

If the embed API determines that it is running in a sandbox, it sets the target origin for the initial postMessage to '*' (otherwise a failure will occur on the postMessage).

New helper methods have been added to the embed builder to determine if the host of the embed is running in sandboxed iframe