Add git workflow to auto gen docs on master push
Npm run docs should be called everytime there's a change
to our source or README, or else the docs will be outdated.
Added a git workflow to auto gen docs on pushes to master
so that manual commits are cleaner with only necessary changes,
and auto commits can contain the auto generated doc changes that
we should not worry about.