feat: Add support for dashboard:edit event (#113)
Adds a new method to the dashboard client that allows an embed host
to set an embedded dashboard into edit mode. Two new events have
been added:
1. dashboard:edit:start
2. dashboard:edit:cancel

Note that the existing event, dashboard:save:complete is sent when a
dashboard save completes.

An additional property has been added to dashboard events that indicate
if the user can actually edit the dashboard.

Documentation has been updated to reflect new events. The dependbot
fix updated typedoc. Some of the types had to be explicitly exported
in order for the same amount of information to be provided. Also, the
run command for typedoc had to be adjusted for the new version.