fix: fix initial iframe creation single thread (#119)
Fixes an issue where the login from the first IFRAME was not
being single threaded. Essentially logins were happening concurrently.
This was not apparent as the test implementation in the server
handled it and insured that all IFRAMEs joined the same session.
This cannot be guaranteed in a customers environment though.