fix: session not found after dev server recycle (#120)
Sessions are cached in memory in the dev server. If the dev
server is recycled the session data is lost. If a generate token
request is received and the session is not found, it now returns
a session time to live of 0. This causes the UI to display the
session expired dialog which is the best result for this scenario,