feat: cookieless embed server cookies
Stores embed session data in the embed servers cookie. Cookie is encrypted
and signed. Replaces in memory cache implementation. The cookie has the
following advantage:
1. can be used in clustered environment.
2. survives server restart.

In addition embed routes used by the development server have been extracted
to there own typescript file. This is in preparation for creating a tsnode
implementation of the demo server.