feat: cookieless embed (#114)
* feat: cookieless embed

Adds support for cookieless embed to the embed-sdk. This is an additive
change and does not impact existing functionality. The actual change to
the SDK is quite small. Note that cookieless embed does require the use
of new endpoints. Examples of using these endpoints have been added to
web development server.

The README has been updated to specify how to utilize cookieless embed.

The larger part of this change is to the demo application. The demo application
has been substantially modified to make it easier to try out embed in
various scenarios. A summary of the changes are as follows:
1. ids of objects to embed can now be specified in the .env file.
   this means it is no longer necessary to customize demo_config.ts.
2. how the embed application behaves can be changed through the demo
   UI. These changes are saved to local storage so survive a page
   reload.
3. The dashboard embed can be hidden/shown. This simulates how a customer
   might create additional IFRAMEs.
4, The look embed can be hidden/shown.
5. The explore embed can be hidden/shown.
6. The extension embed can be hidden/shown.
7. Fonts and components have been made a little prettier (not much but
   it is a little better).
8. The layout is a little reactive. On large monitors, the embeds may be
   displayed side by side.
9. Dashboards can be dynamically resized.
10. Dashboard navigation can be dynamically prevented allowed.
11. Cookieless embed can be dynamically turned on/off.
12. Status messages are now updated for more events.
13. A example of the using the embed postMessage API has been added. Note
    that this is NOT the embed SDK but was added to ensure the cookieless
    embed also works with postMessage.