feat: improve tile dialog rendering support
The Embed SDK and Looker now send and receive information that allows
dialogs opened from tiles to be rendered in a user friendly manner.

1. The top of tile drilling dialogs is scrolled into view.
2. Alert and download dialogs remain in view.

The developer has to make a couple of simple calls to activate this
functionality.

In addition, a convenience method has been added to simplify IFRAME
dynamic height support.