chore: replace pre-commit with husky (#208)
pre-commit is a tool that installs a pre commit hook to
verify the project before the commit takes place. Unfortunately
pre-commit was last published 8 years ago and has a number
of security issues published against.

This change replaces it with husky which does the same thing
but is actively maintained.