feat: Allow url parameters to used when loading using an id
The load functions (except loadQuery) now accept an object
instead of an id as the first parameter. The object can
set the id (required), pushHistory (optional), options
(optional) and parameters (optional). Parameters are appended
as query string parameters. This allows themes to be set
when using the loadFunctions. Filters can also be set if needed.

The single page demo has been updated to demo the use of themes.